fix: rename folder hooks to utils
